requirement.txt
tweepy
logger
nltk-corpus
matplotlib
